Provide means for complete cleanup of LibAwsIO default resources
Creates a mutable struct wrapper for each default resource so
we can add appropraite finalizers. Also adds a convenience close_
method for each resource that will finalize the current resource and
reset it to "null".

This machinery is necessary for precompile workloads to appropriately
do a total cleanup of all process resources (as discovered in
https://github.com/JuliaWeb/HTTP.jl/pull/1213)